refactoring for comment and link the new feature functionality to the main
