+++ main-service/src/main/java/ru/practicum/ewm/main/comment/controller/PrivateCommentController.java
@@ -1,0 +1,43 @@
+package ru.practicum.ewm.main.comment.controller;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.ewm.main.comment.dto.CommentFullDto;
+import ru.practicum.ewm.main.comment.dto.NewCommentDto;
+import ru.practicum.ewm.main.comment.dto.UpdateCommentDto;
+import ru.practicum.ewm.main.comment.service.CommentService;
+
+@RestController
+@RequestMapping("/users/{userId}/events/{eventId}/comments")
+@RequiredArgsConstructor
+public class PrivateCommentController {
+
+    private final CommentService commentService;
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public CommentFullDto addComment(@PathVariable long userId,
+                                     @PathVariable long eventId,
+                                     @Valid @RequestBody NewCommentDto newCommentDto) {
+        return commentService.addComment(userId, eventId, newCommentDto);
+    }
+
+    @PatchMapping("/{commentId}")
+    @ResponseStatus(HttpStatus.OK)
+    public CommentFullDto updateComment(@PathVariable long userId,
+                                        @PathVariable long eventId,
+                                        @PathVariable long commentId,
+                                        @Valid @RequestBody UpdateCommentDto updateCommentDto) {
+        return commentService.updateComment(userId, eventId, commentId, updateCommentDto);
+    }
+
+    @DeleteMapping("/{commentId}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteComment(@PathVariable long userId,
+                              @PathVariable long eventId,
+                              @PathVariable long commentId) {
+        commentService.deleteComment(userId, eventId, commentId);
+    }
+}
