+++ main-service/src/main/java/ru/practicum/ewm/main/comment/controller/PublicCommentController.java
@@ -1,0 +1,33 @@
+package ru.practicum.ewm.main.comment.controller;
+
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.ewm.main.comment.dto.CommentFullDto;
+import ru.practicum.ewm.main.comment.service.CommentService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/comments")
+@RequiredArgsConstructor
+public class PublicCommentController {
+
+    private final CommentService commentService;
+
+    @GetMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public CommentFullDto findComment(@PathVariable long id) {
+        return commentService.findComment(id);
+    }
+
+    @GetMapping("/events/{eventId}")
+    @ResponseStatus(HttpStatus.OK)
+    public List<CommentFullDto> findAllEventComments(@PathVariable long eventId,
+                                                     @RequestParam(defaultValue = "0") @PositiveOrZero int from,
+                                                     @RequestParam(defaultValue = "10") @Positive int size) {
+        return commentService.findAllEventComments(eventId, from, size);
+    }
+}
