+++ main-service/src/main/java/ru/practicum/ewm/main/comment/dto/NewCommentDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.main.event.dto;
+package ru.practicum.ewm.main.comment.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -10,7 +10,7 @@ import lombok.experimental.FieldDefaults;
 @AllArgsConstructor
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class UpdateCommentDto {
+public class NewCommentDto {
     @NotBlank(message = "Field 'text' cannot be null, empty or blank")
     @Size(min = 2, max = 1000, message = "Length of field 'annotation' should be in the range from 2 to 1000")
     String text;
