+++ main-service/src/main/java/ru/practicum/ewm/main/comment/model/Comment.java
@@ -1,10 +1,11 @@
-package ru.practicum.ewm.main.event.model;
+package ru.practicum.ewm.main.comment.model;
 
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+import ru.practicum.ewm.main.event.model.Event;
 import ru.practicum.ewm.main.user.model.User;
 
 import java.time.LocalDateTime;
