+++ main-service/src/main/java/ru/practicum/ewm/main/comment/repository/CommentRepository.java
@@ -1,8 +1,8 @@
-package ru.practicum.ewm.main.event.repository;
+package ru.practicum.ewm.main.comment.repository;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.ewm.main.event.model.Comment;
+import ru.practicum.ewm.main.comment.model.Comment;
 
 import java.util.List;
 
@@ -10,4 +10,6 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
     List<Comment> findAllByEventId(long eventId, PageRequest page);
 
     void deleteAllByEventId(long eventId);
+
+    long countByEventId(long eventId);
 }
