+++ main-service/src/main/java/ru/practicum/ewm/main/event/dto/EventFullDto.java
@@ -43,4 +43,5 @@ public class EventFullDto {
     State state;
     String title;
     long views;
+    long comments;
 }
