+++ main-service/src/main/java/ru/practicum/ewm/main/event/service/EventService.java
@@ -20,20 +20,4 @@ public interface EventService {
     List<EventShortDto> findAllEventsPublic(FindAllEventsPublicParamEntity findAllEventsPublicParamEntity);
 
     EventFullDto findEvent(long eventId);
-
-    CommentFullDto addComment(long userId, long eventId, NewCommentDto newCommentDto);
-
-    CommentFullDto updateComment(long userId, long eventId, long commentId, UpdateCommentDto updateCommentDto);
-
-    CommentFullDto updateCommentAdmin(long commentId, UpdateCommentDto updateCommentDto);
-
-    CommentFullDto findComment(long commentId);
-
-    List<CommentFullDto> findAllEventComments(long eventId, int from, int size);
-
-    void deleteCommentAdmin(long commentId);
-
-    void deleteComment(long userId, long eventId, long commentId);
-
-    void deleteAllEventCommentsAdmin(long eventId);
 }
