+++ main-service/src/main/java/ru/practicum/ewm/main/event/service/EventServiceImpl.java
@@ -10,12 +10,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.main.category.model.Category;
 import ru.practicum.ewm.main.category.repository.CategoryRepository;
+import ru.practicum.ewm.main.comment.repository.CommentRepository;
 import ru.practicum.ewm.main.event.dto.*;
-import ru.practicum.ewm.main.event.mapper.CommentMapper;
 import ru.practicum.ewm.main.event.mapper.EventMapper;
 import ru.practicum.ewm.main.event.mapper.LocationMapper;
 import ru.practicum.ewm.main.event.model.*;
-import ru.practicum.ewm.main.event.repository.CommentRepository;
 import ru.practicum.ewm.main.event.repository.EventRepository;
 import ru.practicum.ewm.main.event.repository.LocationRepository;
 import ru.practicum.ewm.main.exception.type.BadRequestException;
@@ -44,7 +43,7 @@ public class EventServiceImpl implements EventService {
     CategoryRepository categoryRepository;
     StatClient statClient;
     RequestRepository requestRepository;
-    private final CommentRepository commentRepository;
+    CommentRepository commentRepository;
 
     @Override
     public EventFullDto addEvent(long userId, NewEventDto newEventDto) {
@@ -194,79 +193,6 @@ public class EventServiceImpl implements EventService {
         return loadStatisticAndRequest(EventMapper.mapToEventFullDto(event));
     }
 
-    @Override
-    public CommentFullDto addComment(long userId, long eventId, NewCommentDto newCommentDto) {
-        User user = receiveUser(userId);
-        Event event = receiveEvent(eventId);
-        checkValidEventStatusAndRequester(event, user);
-        return CommentMapper.mapToCommentFullDto(commentRepository.save(CommentMapper.mapToComment(event, user, newCommentDto)));
-    }
-
-    @Override
-    public CommentFullDto updateComment(long userId, long eventId, long commentId, UpdateCommentDto updateCommentDto) {
-        User user = receiveUser(userId);
-        Event event = receiveEvent(eventId);
-        Comment comment = receiveComment(commentId);
-        checkValidRequester(event, user, comment);
-        comment.setText(updateCommentDto.getText());
-        return CommentMapper.mapToCommentFullDto(commentRepository.save(comment));
-    }
-
-    @Override
-    public CommentFullDto updateCommentAdmin(long commentId, UpdateCommentDto updateCommentDto) {
-        Comment comment = receiveComment(commentId);
-        comment.setText(updateCommentDto.getText());
-        return CommentMapper.mapToCommentFullDto(commentRepository.save(comment));
-    }
-
-    @Override
-    public CommentFullDto findComment(long commentId) {
-        return CommentMapper.mapToCommentFullDto(receiveComment(commentId));
-    }
-
-    @Override
-    public List<CommentFullDto> findAllEventComments(long eventId, int from, int size) {
-        PageRequest page = PageRequest.of(from, size);
-        return commentRepository.findAllByEventId(eventId, page).stream().map(CommentMapper::mapToCommentFullDto).toList();
-    }
-
-    @Override
-    public void deleteCommentAdmin(long commentId) {
-        commentRepository.deleteById(commentId);
-    }
-
-    @Override
-    public void deleteComment(long userId, long eventId, long commentId) {
-        User user = receiveUser(userId);
-        Event event = receiveEvent(eventId);
-        Comment comment = receiveComment(commentId);
-        checkValidRequester(event, user, comment);
-        commentRepository.deleteById(commentId);
-    }
-
-    @Override
-    public void deleteAllEventCommentsAdmin(long eventId) {
-        commentRepository.deleteAllByEventId(eventId);
-    }
-
-    private void checkValidEventStatusAndRequester(Event event, User user) {
-        if (event.getState() != State.PUBLISHED) {
-            throw new BadRequestException("Event with id=" + event.getId() + " must be published");
-        }
-        if (event.isRequestModeration() || event.getParticipantLimit() != 0) {
-            if (!event.getInitiator().equals(user) && requestRepository.findByRequesterIdAndEventId(user.getId(), event.getId())
-                    .filter(o -> o.getStatus() == Status.CONFIRMED).isEmpty()) {
-                throw new BadRequestException("User with id=" + user.getId() + " cannot work with comments");
-            }
-        }
-    }
-
-    private void checkValidRequester(Event event, User user, Comment comment) {
-        if (!comment.getAuthor().equals(user) && !event.getInitiator().equals(user)) {
-            throw new BadRequestException("User " + user.getId() + " cannot delete a comment " + comment.getId() + " that is not his own.");
-        }
-    }
-
     private void updateFields(Event event, UpdateEventFieldsEntity updateEventFieldsEntity) {
         if (updateEventFieldsEntity.hasEventDate()) {
             if (updateEventFieldsEntity.getEventDate().isBefore(LocalDateTime.now().plusHours(2))) {
@@ -307,7 +233,8 @@ public class EventServiceImpl implements EventService {
         long amountOfViews = statClient.getStat(event.getCreated(), LocalDateTime.now(), List.of("/events/" + event.getId()), true).stream()
                 .map(ViewStatsDto::getHits)
                 .reduce(0L, Long::sum);
-        return EventMapper.mapToEventShortDto(event, amountOfConfirmedRequests, amountOfViews);
+        long amountOfComments = commentRepository.countByEventId(event.getId());
+        return EventMapper.mapToEventShortDto(event, amountOfConfirmedRequests, amountOfViews, amountOfComments);
     }
 
     private EventFullDto loadStatisticAndRequest(EventFullDto event) {
@@ -317,6 +244,8 @@ public class EventServiceImpl implements EventService {
                 .map(ViewStatsDto::getHits)
                 .reduce(0L, Long::sum);
         event.setViews(amountOfViews);
+        long amountOfComments = commentRepository.countByEventId(event.getId());
+        event.setComments(amountOfComments);
         return event;
     }
 
@@ -343,6 +272,7 @@ public class EventServiceImpl implements EventService {
                         .filter(view -> view.getUri().equals("/events/" + event.getId()))
                         .map(ViewStatsDto::getHits)
                         .reduce(0L, Long::sum)))
+                .peek(event -> event.setComments(commentRepository.countByEventId(event.getId())))
                 .toList();
     }
 
@@ -405,11 +335,6 @@ public class EventServiceImpl implements EventService {
     private Event receiveEvent(long eventId) {
         return eventRepository.findById(eventId)
                 .orElseThrow(() -> new NotFoundException("Event with id=" + eventId + " was not found"));
-    }
-
-    private Comment receiveComment(long commentId) {
-        return commentRepository.findById(commentId)
-                .orElseThrow(() -> new NotFoundException("Comment with id=" + commentId + " was not found"));
     }
 
     private void checkValidUserForEvent(User user, Event event) {
